backend/requirements.txt
django>=5.1.6
djangorestframework>=3.15.2
django-cors-headers>=4.4.0
djangorestframework-simplejwt>=5.3.1
PyJWT>=2.8.0
mysqlclient>=2.2.4
python-decouple>=3.8
whitenoise>=6.7.0
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.5.0
joblib>=1.4.2
pillow>=10.4.0
django-cleanup>=8.0.0
gunicorn>=23.0.0
